fix: Added web link for desktop error page
