+++ src/pages/GenericErrorPage/ErrorBodyText/index.website.js
@@ -1,0 +1,1 @@
+export default () => null;
